cargo.eclass: add multi-target support to cargo
Start to use rust eclass into cargo eclass in order to support multi-target
binaries installation.

If an ebuild is compiled with multiple rust targets the binaries will be
installed into /usr/lib/${rust-target}/bin. Is not the perfect solution to
everyone to have some binaries outside $PATH, but is more clear to have a
single version of the binary handled by eselect into /usr/bin instead of
multiple versions. This because if the binary needs to be linked to rustc
library could be stand to a conflict.

Signed-off-by: gibix <gibix@riseup.net>